Реализовал добавление заказа
Заказ добавляется следующим образом: создается кортеж в отношении 'orders'. В этом кортеже есть атрибут 'order_id'. Это значение вставляется во все кортежи отношения 'dish_basket', у которых 'user_id' - айди текущего пользователя и 'order_id' IS NULL, то есть, это блюдо заказано ещё не было, оно находится пока что просто в корзине. Наличие не NULL 'order_id' будет означать, что блюдо было заказано и относится к заказу, id которого равен 'order_id'

+++ api/basket/helpers/get_dish_ids.php
@@ -19,7 +19,7 @@ function showCartByDishIDs($dishesInCart)
             foreach ($dishesInCart as $value) {
                 $currentDishAmount = $Link->query("SELECT amount FROM dish_basket WHERE user_id = '$userID' AND dish_id = '$value[0]'")->fetch_assoc()['amount'];
                 $currentDish = $Link->query("SELECT name, price, image, dish_id FROM dish WHERE dish_id = '$value[0]'")->fetch_assoc();
-                $cart[sizeof($cart)] = array("name" => $currentDish['name'], "price" => (int)$currentDish['price'], "totalPrice" => (int)$currentDish['price'] * (int)$currentDishAmount, "amount" => (int)$currentDishAmount, "image" => $currentDish['image'], "id" => $value[0]);
+                $cart[sizeof($cart)] = array("name" => $currentDish['name'], "price" => (float)$currentDish['price'], "totalPrice" => (float)$currentDish['price'] * (int)$currentDishAmount, "amount" => (int)$currentDishAmount, "image" => $currentDish['image'], "id" => $value[0]);
             }
             echo json_encode($cart);
